feat: Display Bookshelves and Books Dynamically
Load Books via the BooksAPI then display them on the page
dynamically, replacing the static content.

Given we know the Bookshelves in use we can hard-code these
in the component and iterate over them, simplifying the filter
logic by keeping it all in one place.

An alternative would be to dynamically fetch the unique list of
bookshelves attached to books and use those instead of a hard-
coded object/array.